Refactor into separate modules & put tests back
