+++ Cargo.toml
@@ -30,6 +30,7 @@ libarchive3-sys = "0.1.2"
 
 [dev-dependencies]
 flate2 = "0.2.14"
+byteorder = "0.5"
 lazy_static = "0.2.1"
 
 [build-dependencies]
